requirements.txt
numpy==1.23.0
Pillow==10.0.1
pygame==2.1.2
